+++ planning/planning_test_utils/CMakeLists.txt
@@ -4,19 +4,19 @@ project(planning_test_utils)
 find_package(autoware_cmake REQUIRED)
 autoware_package()
 
-ament_auto_add_library(route_parser SHARED
-  src/route_parser.cpp)
+ament_auto_add_library(mock_data_parser SHARED
+  src/mock_data_parser.cpp)
 
-target_link_libraries(route_parser
+target_link_libraries(mock_data_parser
   yaml-cpp
 )
 
 if(BUILD_TESTING)
-  ament_add_ros_isolated_gtest(test_route_parser
-    test/test_route_parser.cpp)
+  ament_add_ros_isolated_gtest(test_mock_data_parser
+    test/test_mock_data_parser.cpp)
 
-  target_link_libraries(test_route_parser
-    route_parser)
+  target_link_libraries(test_mock_data_parser
+    mock_data_parser)
 endif()
 
 ament_auto_package(INSTALL_TO_SHARE
